+++ NEWS.md
@@ -1,3 +1,10 @@
+# pegboard 0.0.0.9016
+
+ - the Episode class now has the `$headings` active binding and the 
+   `$validate_headings()` method to validate headings.
+ - We now use {testthat} 3e for our tests
+ - Cloning an Episode object is now more reliable
+
 # pegboard 0.0.0.9015
 
  - the Episode class is now a sub-class of the tinkr::yarn class and gains the